monday_to_salesforce/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
python-dotenv
mangum